Created new landing page that give user the option to choose retail or rental
